+++ bun/cmd/import.go
@@ -1,15 +1,9 @@
 package cmd
 
 import (
-	_ "github.com/adyatlov/bun/check/dcosversion"
-	_ "github.com/adyatlov/bun/check/health"
-	_ "github.com/adyatlov/bun/check/marathon/deployments"
-	_ "github.com/adyatlov/bun/check/mesos/actormailboxes"
-	_ "github.com/adyatlov/bun/check/mesos/unmountvolume"
-	_ "github.com/adyatlov/bun/check/nodecount"
-	_ "github.com/adyatlov/bun/file/dcosversionfile"
-	_ "github.com/adyatlov/bun/file/healthfile"
-	_ "github.com/adyatlov/bun/file/marathon/deploymentsfile"
-	_ "github.com/adyatlov/bun/file/mesos/actormailboxesfile"
-	_ "github.com/adyatlov/bun/file/mesos/agentlogfile"
+	_ "github.com/adyatlov/bun/checks/dcosversion"
+	_ "github.com/adyatlov/bun/checks/health"
+	_ "github.com/adyatlov/bun/checks/marathon/deployments"
+	_ "github.com/adyatlov/bun/checks/mesos/actormailboxes"
+	_ "github.com/adyatlov/bun/checks/nodecount"
 )