+++ bun/main.go
@@ -4,9 +4,6 @@ import (
 	"github.com/adyatlov/bun/bun/cmd"
 )
 
-const printProgress = false
-const printLong = false
-
 func main() {
 	cmd.Execute()
 }