+++ checks/dcosversion/check.go
@@ -2,9 +2,8 @@ package dcosversion
 
 import (
 	"fmt"
-
 	"github.com/adyatlov/bun"
-	"github.com/adyatlov/bun/file/dcosversionfile"
+	"github.com/adyatlov/bun/filetypes"
 )
 
 func init() {
@@ -12,16 +11,17 @@ func init() {
 		Name: "dcos-version",
 		Description: "Verify that all hosts in the cluster have the " +
 			"same DC/OS version installed",
-		ForEachMaster:      check,
-		ForEachAgent:       check,
-		ForEachPublicAgent: check,
-		Interpret:          interpret,
+		CollectFromMasters:      collect,
+		CollectFromAgents:       collect,
+		CollectFromPublicAgents: collect,
+		Aggregate:               aggregate,
 	}
-	builder.BuildAndRegister()
+	check := builder.Build()
+	bun.RegisterCheck(check)
 }
 
-func check(host bun.Host) (ok bool, details interface{}, err error) {
-	v := dcosversionfile.Version{}
+func collect(host bun.Host) (ok bool, details interface{}, err error) {
+	v := filetypes.Version{}
 	if err = host.ReadJSON("dcos-version", &v); err != nil {
 		return
 	}
@@ -30,7 +30,7 @@ func check(host bun.Host) (ok bool, details interface{}, err error) {
 	return
 }
 
-func interpret(c *bun.Check, b bun.CheckBuilder) {
+func aggregate(c *bun.Check, b bun.CheckBuilder) {
 	version := ""
 	// Compare versions
 	details := []string{}
@@ -39,7 +39,6 @@ func interpret(c *bun.Check, b bun.CheckBuilder) {
 		v := r.Details.(string)
 		if version == "" {
 			version = v
-			continue
 		}
 		if v != version {
 			ok = false
@@ -49,6 +48,7 @@ func interpret(c *bun.Check, b bun.CheckBuilder) {
 	}
 	// No need to interpret problems, as we didn't create it in the host check.
 	if ok {
+		c.OKs = details
 		c.Summary = fmt.Sprintf("All versions are the same: %v.", version)
 	} else {
 		c.Problems = details