+++ checks/search_checks_yaml.go
@@ -1,0 +1,12 @@
+package checks
+
+const searchChecksYAML = `
+- name: unmount-volume
+  description: Checks if Mesos agents had problems unmounting local persistent volumes. MESOS-8830
+  fileTypeName: mesos-agent-log
+  searchString: Failed to remove rootfs mount point
+- name: disk-space-exhibitor
+  description: Check disk space errors in Exhibitor logs
+  fileTypeName: exhibitor-log
+  searchString: Failed to remove rootfs mount point
+`