+++ file_type.go
@@ -9,23 +9,27 @@ import (
 type ContentType string
 
 const (
-	// JSON represents JSON files.
-	JSON ContentType = "JSON"
-	// Journal represents Journal files.
-	Journal = "journal"
-	// Dmesg represents dmesg files.
-	Dmesg = "dmesg"
+	// CTJson represents CTJson files.
+	CTJson ContentType = "JSON"
+	// CTJournal represents CTJournal files.
+	CTJournal = "journal"
+	// CTDmesg represents dmesg files.
+	CTDmesg = "dmesg"
+	// CTOutput is a output of a command.
+	CTOutput = "output"
+	//CTOther file types
+	CTOther = "other"
 )
 
 // FileType Describes a kind of files in the bundle (e.g. dcos-marathon.service).
 type FileType struct {
-	Name        string
-	ContentType ContentType
-	Paths       []string
-	Description string
+	Name        string `yaml:"name"`
+	ContentType ContentType `yaml:"contentType"`
+	Paths       []string `yaml:"paths"`
+	Description string `yaml:"description"`
 	// DirTypes defines on which host types this file can be found.
 	// For example, dcos-marathon.service file can be found only on the masters.
-	DirTypes []DirType
+	DirTypes []DirType `yaml:"dirTypes"`
 }
 
 var (